// src/model/register.ts
export class IdValuePair {
    id: string;
    value: string;
}
export class RegistrationDetails {
    role: number;
    id: number;
    firstName: string;
    lastName: string;
    dateOfBirth: string;
    gender: number;
    mobileNumber: string;
    email: string;
    password: string;
}

export class EmpListData {
    id: number;
    role: number;
    firstName: string;
    lastName: string;
    email: string;
}

export class UsedID {
    id: number;
}